fix: fix template cli and use raw jinja2

+++ repo_scaffold/templates/template-python/hooks/post_gen_project.py
@@ -12,24 +12,10 @@ def remove_cli():
 
 def remove_github_actions():
     """Remove GitHub Actions configuration if not needed."""
-    github_dir = os.path.join("{{cookiecutter.project_slug}}", ".github")
-    if os.path.exists(github_dir):
-        shutil.rmtree(github_dir)
-
-
-def remove_mkdocs():
-    """Remove MkDocs related files if not needed."""
-    files_to_remove = [
-        "mkdocs.yml",
-        os.path.join("docs"),
-    ]
-    for file in files_to_remove:
-        path = os.path.join("{{cookiecutter.project_slug}}", file)
-        if os.path.exists(path):
-            if os.path.isdir(path):
-                shutil.rmtree(path)
-            else:
-                os.remove(path)
+    if "{{cookiecutter.use_github_actions}}" == "no":
+        github_dir = os.path.join(".github")
+        if os.path.exists(github_dir):
+            shutil.rmtree(github_dir)
 
 
 def init_project_depends():
@@ -39,10 +25,6 @@ def init_project_depends():
     
     # 安装基础开发依赖
     subprocess.run(["uv", "sync", "--extra", "dev"], check=True)
-    
-    # 如果启用了文档功能，安装文档依赖
-    if "{{cookiecutter.use_mkdocs}}" == "yes":
-        subprocess.run(["uv", "sync", "--extra", "docs"], check=True)
 
 
 if __name__ == "__main__":
@@ -51,8 +33,5 @@ if __name__ == "__main__":
         
     if "{{cookiecutter.use_github_actions}}" == "no":
         remove_github_actions()
-        
-    if "{{cookiecutter.use_mkdocs}}" == "no":
-        remove_mkdocs()
         
     init_project_depends()
